refactored displayMices
